requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attr==0.3.1
attrs==23.1.0
azure-core==1.26.4
azure-storage-blob==12.16.0
beautifulsoup4==4.12.2
bleach==5.0.1
blinker==1.6.2
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
Brlapi==0.8.3
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.0
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
Django==3.2.16
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
expiringdict==1.2.2
eyeD3==0.8.10
filelock==3.6.0
Flask==2.3.2
Flask-Caching==2.0.2
flask-restx==1.1.0
flatbuffers==23.5.9
fsspec==2023.5.0
gast==0.4.0
google-api-core==2.10.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.3
grpcio==1.54.0
grpcio-status==1.48.2
h5py==3.8.0
htmlmin==0.1.12
httplib2==0.20.2
huggingface-hub==0.14.1
idna==3.4
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
jax==0.4.10
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==3.2.0
keras==2.12.0
keyring==23.5.0
label-studio==1.7.3
label-studio-converter==0.0.51
label-studio-tools==0.0.2
launchdarkly-server-sdk==7.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.2
ml-dtypes==0.1.0
more-itertools==8.10.0
nemo-emblems==5.6.0
netaddr==0.8.0
netifaces==0.11.0
nltk==3.6.7
numpy==1.23.5
oauthlib==3.2.0
onboard==1.4.1
openapi-codec==1.3.2
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==23.1
PAM==0.4.2
pandas==1.3.5
pexpect==4.8.0
Pillow==9.3.0
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pydantic==1.10.7
pyelftools==0.27
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pyrsistent==0.19.3
pysmbc==1.0.23
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-json-logger==2.0.4
python-magic==0.4.24
python-xlib==0.29
pytz==2019.3
pyxdg==0.27
PyYAML==5.4.1
redis==4.5.4
regex==2023.5.5
reportlab==3.6.8
requests==2.28.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
scipy==1.10.1
SecretStorage==3.3.1
semver==2.13.0
sentry-sdk==1.22.1
setproctitle==1.2.2
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
systemd-python==234
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.5.0
ua-parser==0.16.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.7.0
Unidecode==1.3.3
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.3.3
wrapt==1.14.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
xmljson==0.2.0
youtube-dl==2021.12.17
zipp==1.0.0
